Refactor: Move component code to __init__.py
I moved all Python code from lgtv_uart_external/component.py
to lgtv_uart_external/__init__.py. The file component.py was subsequently deleted.

This change centralizes the component's Python code into the standard
__init__.py file, simplifying the directory structure for the
ESPHome custom component.
